Adds a guard around instance termination.
[#119028263](https://www.pivotaltracker.com/story/show/119028263)

+++ ci/tasks/ensure-terminated.sh
@@ -15,5 +15,8 @@ if [ ! -z "${vpc_id}" ] ; then
   instances=$(aws ec2 describe-instances --query "Reservations[*].Instances[*].InstanceId[]" --filters "Name=vpc-id,Values=${vpc_id}" | jq '.[]' --raw-output)
   instance_list=$(echo ${instances} | sed "s/[\n\r]+/ /g")
 
-  aws ec2 terminate-instances --instance-ids ${instance_list}
+  # if it's not an empty string (of any length)...
+  if [ ! -z "${instance_list// }" ] ; then
+    aws ec2 terminate-instances --instance-ids ${instance_list}
+  fi
 fi
